aios/vision_desktop/requirements.txt
PyQt5
pillow
pyautogui
transformers
torch
